Changed FanSpin.cs so that the fan can spin slower (if need be)

+++ Assets/Scripts/FanSpin.cs
@@ -6,7 +6,7 @@ public class FanSpin : MonoBehaviour
 {
 
     [SerializeField]
-    int spinSpeed = 20;
+    float spinSpeed = 20f;
 
     void Start()
     {
